Fits a mixed regression model using Stan It requires the same number of observation/locations per day. It requires location id (locID) and points id (pid). The locID are unique for each site. The pid is unique for each observation. Missing values are allowed in the response but not in the covariates.

Fits a mixed regression model using Stan It requires the same number of observation/locations per day. It requires location id (locID) and points id (pid). The locID are unique for each site. The pid is unique for each observation. Missing values are allowed in the response but not in the covariates.

ssnbayes(
  formula = formula,
  data = data,
  path = path,
  time_method = time_method,
  space_method = space_method,
  iter = 3000,
  warmup = 1500,
  chains = 3,
  refresh = max(iter/100, 1),
  net = 1,
  addfunccol = addfunccol,
  loglik = F,
  seed = seed
)

Arguments

formula
A formula as in lm()

data
A long data frame containing the locations, dates, covariates and the response variable. It has to have the locID and date. No missing values are allowed in the covariates.

path
Path with the name of the SSN object

time_method
A list specifying the temporal structure (ar = Autorregressive; var = Vector autorregression) and coumn in the data with the time variable.

space_method
A list defining if use or not of an SSN object and the spatial correlation structure. The second element is the spatial covariance structure. A 3rd element is a a list with the lon and lat for Euclidean distance models.

iter
Number of iterations

warmup
Warm up samples

chains
Number of chains

refresh
Sampler refreshing rate

net
The network id (optional). Used when the SSN object cotains multiple networks.

addfunccol
Variable to compute the additive function. Used to compute the spatial weights.

loglik
Logic parameter denoting if the loglik will be computed by the model.

seed
(optional) A seed for reproducibility

Value

A list with the model fit

Examples

#fit_td <- ssnbayes(formula = 'y ~ X1 + X2 + X3',
#                    data = data, ssn = ssn,
#                    space_method = list("use_ssn", "Exponential.taildown"),
#                    time_method = list("ar", "date"),
#                    iter = 3000,
#                    warmup = 1500,
#                    chains = 3)